Restructuring metadata test for maintainability
