+++ locksmith/__tests__/controllers/metadataController/updateKeyMetadata.test.ts
@@ -1,0 +1,93 @@
+import request from 'supertest'
+import * as sigUtil from 'eth-sig-util'
+import * as ethJsUtil from 'ethereumjs-util'
+
+const app = require('../../../src/app')
+const Base64 = require('../../../src/utils/base64')
+
+function generateKeyTypedData(message: any) {
+  return {
+    types: {
+      EIP712Domain: [
+        { name: 'name', type: 'string' },
+        { name: 'version', type: 'string' },
+        { name: 'chainId', type: 'uint256' },
+        { name: 'verifyingContract', type: 'address' },
+        { name: 'salt', type: 'bytes32' },
+      ],
+      KeyMetadata: [],
+    },
+    domain: {
+      name: 'Unlock',
+      version: '1',
+    },
+    primaryType: 'KeyMetadata',
+    message,
+  }
+}
+
+const privateKey = ethJsUtil.toBuffer(
+  '0xfd8abdd241b9e7679e3ef88f05b31545816d6fbcaf11e86ebd5a57ba281ce229'
+)
+
+const mockOnChainLockOwnership = {
+  owner: jest.fn(() => {
+    return Promise.resolve('0xAaAdEED4c0B861cB36f4cE006a9C90BA2E43fdc2')
+  }),
+}
+
+jest.mock('../../../src/utils/lockData', () => {
+  return function() {
+    return mockOnChainLockOwnership
+  }
+})
+
+const mockKeyHoldersByLock = {
+  getKeyHoldingAddresses: jest.fn(() => {
+    return Promise.resolve(['0xAaAdEED4c0B861cB36f4cE006a9C90BA2E43fdc2'])
+  }),
+}
+
+jest.mock('../../../src/graphql/datasource/keyholdersByLock', () => ({
+  __esModule: true,
+  KeyHoldersByLock: jest.fn(() => {
+    return mockKeyHoldersByLock
+  }),
+}))
+
+describe('updateKeyMetadata', () => {
+  describe('when the signee does not own the lock', () => {
+    let typedData: any
+    beforeAll(() => {
+      typedData = generateKeyTypedData({
+        KeyMetaData: {
+          custom_field: 'custom value',
+          owner: '0xaaadeed4c0b861cb36f4ce006a9c90ba2e43fdc2',
+        },
+      })
+
+      mockOnChainLockOwnership.owner = jest.fn(() => {
+        return Promise.resolve(
+          '0xAaAdEED4c0B861cB36f4cE006a9C90BA2E43fdfegdrtfeghr'
+        )
+      })
+    })
+
+    it('returns unauthorized', async () => {
+      expect.assertions(1)
+
+      const sig = sigUtil.signTypedData(privateKey, {
+        data: typedData,
+        from: '',
+      })
+
+      const response = await request(app)
+        .put('/api/key/0x95de5F777A3e283bFf0c47374998E10D8A2183C7/5')
+        .set('Accept', 'json')
+        .set('Authorization', `Bearer ${Base64.encode(sig)}`)
+        .send(typedData)
+
+      expect(response.status).toEqual(401)
+    })
+  })
+})
